+++ fabric/src/main/java/com/lx862/mtrscripting/mod/MTRScripting.java
@@ -1,5 +1,7 @@
 package com.lx862.mtrscripting.mod;
 
+import com.lx862.jcm.mod.Constants;
+import com.lx862.mtrscripting.api.ScriptingAPI;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.mtr.mapping.holder.Identifier;
@@ -10,5 +12,9 @@ public class MTRScripting {
 
     public static Identifier id(String path) {
         return new Identifier(MOD_ID, path);
+    }
+
+    public static void init() {
+        ScriptingAPI.registerAddonVersion(MOD_ID, Constants.MOD_VERSION);
     }
 }